Rename typing -> _typing
