+++ tests/molecules.py
@@ -17,7 +17,7 @@ Molecules for testing. Taken from https://github.com/grimme-lab/mstore.
 """
 import torch
 
-from tad_dftd3.typing import Dict, Molecule
+from tad_dftd3._typing import Dict, Molecule
 from tad_dftd3.util import to_number
 
 mols: Dict[str, Molecule] = {
